server: unix 'rwx' permissions for server.access bundle
Each time we allowed access to some path through the server, we would
grant all "rwx" permissions, which may be a bit dangerous sometimes.

For example, we want to grant /sbin/cf-agent access for 'cf-runagent',
but that doesn't mean we want that binary to be remotely writen to!

Or, we may grant access to /var/log/syslog, but not allow a remote party
to erase our logs.

So, have a Unix-like permission system, where read/write/execute
permissions can explicitly be granted or revoked.

By default, all rights are granted, for compatibility with previous files.