server.c: fix sending of large files ( >3GB)
The code there was using 'int' for the stat.st_size, which is 64bit in
some platforms. This meant that the periodic stat() of the file would
be truncated and hence fail, sending a "BAD: file changed" to the remote.
Also, the return value of stat() is now being checked.

Note: this may need further improvement, by ensuring int64 support for
all filesizes.