cf-server: don't keep report_context throughout server process
In an attempt to get rid of cf-server flooding syslog with emergency
messages, discovered that the report_context of cf-server would once
be instantiated during main() , then passed down the fork()'ed child
as open file descriptors and re-used during the lifetime of the proc.

This doesn't seem right. We'd better only open the files, and close
them shortly just before the promises are checked (if ever needed).