verify_packages: support "list of updates" from packaging tool
New feature!

Until now, the only way to promise updates from an external list would
be to feed them as "patches" of the package manager. However, this does
not work well with the "update" package_policy, nor is it consistent
with (not) specifying the package_version in the promise.

Let us do it right this time:
 - every time the packages promise is evaluated, we run the command for
   "package_new_versions". We expect that to be volatile across cf-agent
   runs, so don't save it in a cache file.
   Assumption: the output of the command is same as the "package_list"
   YMMV, we could add more regex-es there.

 - if a package promises an explicit version, we only accept updates that
   /start with/ that ver. eg. package_version => "3.4." will allow update
   to "3.4.0-b1", "3.4.2-fix8" but not "3.5" or even "3.44"

 - then, if an update is available, and arch matches, we use that version
   as if it were the promised one.