verify_package: "update" of a non-installed package should not err
Partial revert of: 49a48dc5fae1064f1b1b0afe (7/4/2010)

IMHO an "update" that doesn't find the package installed shall not be
an error. Instead, it could just display the verbose message, as before.

In fact, the best approach in the long term would be to have an extra
verb, apart from "update" for "update and check" or "maybe update"...