libcfnet: big (65k) buffer support for file transfers
So far, all libcfnet transactions were limited to 4k buffers. File copy
was even using half of that, 2048-byte buffers. This meant that a net
send() was called every 2k, which is low enough to alias on the 1492b
of ethernet MTU.
The result: poor performance and the Ethernet flooding with ACK packets.

As an experiment, we raise the buffer size to the maximum our 6-digit
protocol header can support. 65k seems reasonable for TCP packets, minus
some bytes for our header.

Fat warning: this BREAKS protocol compatibility with older cfengine2/3
agents, because we cannot negotiate between 4k/65k buffers yet. Older
remote pairs may refuse to speak to us.

This patch also only works for unencrypted transfers. The code of SGET
is yet to be written.