cf-execd,libenv: only run cf-execd with minimal classes, don't reload
cf-execd's role is merely to run cf-agent, always, regardless of any
volatile conditions. Re-evaluating classes like the system's name,
architecture or monitor variables are a waste of resources and allow
for an undesired use of "cf-execd".

Also, for all components[1], an early exit when network is not loaded
may not be desired. Try to run with "interfaces" list empty.

[1] except for cf-serverd, perhaps?