libpromises: future syntax support, draft PoC implementation
When deploying the same masterfiles to a herd with mixed CFEngine
versions, these masterfiles shall validate against all "cf-promises".

But this would stop us from using features not found in earlier vers.

Promise bundles can overcome that, because they can be conditionally
included, depending on the CFEngine version

Control bodies cannot. They would have to use only the classic keywords.

With this patch, we can explicitly define a list of keywords that will
pass as valid syntax and be skipped in parsers. libpromises will first
lookup the body/bundle/constraint in the supported ones, and then fall
over to the "future syntax" list.

Future syntax is defined in control body, because this is the first to
ever get parsed. It is a stringlist, because that is the only possible
collection item in that context. These future-syntax strings are free,
meaning that they could have any form (so far, one dot is mandatory).
That is, future syntax is designed to be future-compatible itself, by
allowing strings that the parser will not use at the moment.

So far, only body constraints are covered.

Example input file:

body common control {
    future_syntax => { "body.common.control.bwlimitw",
                    "body.agent.control.foo_bar"};
    bwlimitw => "4M";
}

body agent control {
    foo_bar => "bar";
}