libcfnet: avoid double DNS lookup on CacheServerConnection
One of my hosts did emit this error:
    ... Failed to resolve <my server>
    client_code.c:1223: Programming Error: ServerNotBusy: No connection found!
which really means that an active connection had not entered the
server_list cache properly.

This is explained by the void return path of that function: if the
connection is properly established, and /then/ the local resolver fails
to provide the IP for the policy hub hostname (eg. in round-robin with
a failing nameserver), then that connection will not hit the cache and
thus not be freed properly.

I noticed that we do have the 'conn->remote_ip' as a resolved string,
so it makes sense to re-use that, saving one DNS resolution call.